fix: refactor types
